compose/backend/requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
Werkzeug==3.0.4
wfastcgi==3.0.0
